requirements.txt
geopandas==0.12.2
fastapi==0.95.0
uvicorn==0.21.1
pandas==1.5.3
pyproj==3.4.1
pyshp==2.3.1
shapely==2.0.1
prisma==0.8.2
python-dotenv